s/audit: Move ocsf builders to static methods
- This will be useful in future commits where a type can be passed to a
templated method and static polymorphic dispatch can be used to
determine which builder method to choose to construct the type
